+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and the versioning aim to respect [Semantic Versioning](http://semver.org/spec/v
 
 Here is a template for new release sections
 
+## [Unreleased]
+### Added
+- possibility to create legend from JSON file
+
 ## [3.0.3] - 2025-07-24
 ### Fixed
 - duplicate loading of legend.js